:white_check_mark: Test default color system arg
Test for regression accidentally introduced. Callbacks in Typer are picky if used with Enums, and will return None if the callback returns a type not expected by original parameter type hint. This test attempt to guard the fact that the default value for color system should be "auto" after transformation by the callback.